+++ mail/mail-uniffi/src/mail/session.rs
@@ -692,6 +692,12 @@ impl MailSession {
         &self.mail_ctx
     }
 
+    /// Get the mail context wrapped in [`Arc`]
+    #[must_use]
+    pub fn ctx_arc(&self) -> Arc<MailContext> {
+        Arc::clone(&self.mail_ctx)
+    }
+
     /// Get the session database connection.
     #[must_use]
     pub fn session_stash(&self) -> &Stash {